Initial upload
A prototype IRAF CL script. This was for developing a script to test for what instrument (CCD or any detector) is stored in the keyword INSTRUME. 
Keywords are extracted from image headers, and example keyword is read in, and a logical test determines if the script proceeds or ends. This is useful if there are detector specific reduction procedures to follow.