second upload
Slight changes incorporating what I learned from the test-instrume.cl script. The test for correct CCD in the image header keyword INSTRUME 
determines if the script continues to reduce the data or it ends. This prevents the user from running reductions with specific needs for one detector on data from a different detector.