Coordinate coo-file editing script
Python script to edit an input coo-file created using daofind. This will take an input file and write an output file including the header information, but excluding saturated stars which are selected and removed using the Sharpness statistics. Image regions are also excluded so edge effects may be removed from the daofind results.